# Use from_gen to parse utf* strings.