--- a/.ocamlformat
+++ b/.ocamlformat
@@ -1,4 +1,4 @@
-version=0.24.1
+version=0.25.1
 profile = conventional
 break-separators = after
 space-around-lists = false