--- a/.ocamlformat-ignore
+++ b/.ocamlformat-ignore
@@ -1,1 +1,0 @@
-src/syntax/unicode.ml