--- a/CHANGES.md
+++ b/CHANGES.md
@@ -1,5 +1,7 @@
 # dev
 - Restore compatibility with OCaml 4.08
+- Use `Sedlexing.Utf8.from_gen` to initialize UTF8 lexing buffers from
+  string.
 
 # 3.1:
 - Fix directly nested sedlex matches (@smuenzel, PR #117, fixes: #12)