This PR optimizes memory and CPU usage but using the `from_gen` API to parse strings in `UTF8` and `UTF16` modules.