@hhugo 2023-06-01
Some tests be nice.
Also, what the performance impact of your change for valid input ?