@toots 2023-06-03
Here's a test program:

```ocaml
let read_whole_file filename =
  let ch = open_in_bin filename in
  let s = really_input_string ch (in_channel_length ch) in
  close_in ch;
  s

let () =
  let s = read_whole_file "/tmp/medialib.json" in
  let lexbuf = Sedlexing.Utf8.from_string s in
  ignore lexbuf
```

Tested on a `4.8M` json file:

Before the changes:
```shell
% time ./sedbench
./sedbench  0.06s user 0.01s system 96% cpu 0.080 total
```

After the changes:
```shell
% time ./sedbench
./sedbench  0.00s user 0.00s system 74% cpu 0.004 total
```

There seems to be a difference between cold and warm during testing so these are the result of the second run in each case and it was confirmed that subsequent runs where in the same range.

Overall, it's about 20x faster to use `from_gen` to _initialize_ the lexbuf!

Changelog entry has been added, this PR should be read for merge now!