@hhugo 2023-06-04
Your benchmark doesn't lex anything, it just create the lexbuf. I would like to see the result that include lexing the whole content.