@toots 2023-06-10
> Your benchmark doesn't lex anything, it just create the lexbuf. I would like to see the result that include lexing the whole content.

Since the patch is focused on the lexbuf creation, I would expect the gains to be on the lexbuf creation and anything after that to be additively static.

Indeed:

Without patch:
```
% time ./_build/default/bench.exe
DONE
./_build/default/bench.exe  0.09s user 0.01s system 95% cpu 0.112 total
```

With patch:
```
❯ time ./_build/default/bench.exe
DONE
./_build/default/bench.exe  0.00s user 0.00s system 80% cpu 0.008 total
```

<details>
  <summary>Benchmark code</summary>

```ocaml
open Sedlexing

type token =
  | EOF
  | LBRACKET
  | RBRACKET
  | COLON
  | COMMA
  | MINUS
  | PLUS
  | LSQUARE
  | RSQUARE
  | STRING of string
  | BOOL of bool
  | NUMBER of string

exception LexError of Lexing.position * string

let digit = [%sedlex.regexp? '0' .. '9']
let number = [%sedlex.regexp? Plus digit]
let blank = [%sedlex.regexp? ' ' | '\t']
let newline = [%sedlex.regexp? '\r' | '\n' | "\r\n"]
let any_blank = [%sedlex.regexp? blank | newline]
let letter = [%sedlex.regexp? 'a' .. 'z' | 'A' .. 'Z']
let decimal_ascii = [%sedlex.regexp? Plus '0' .. '9']
let octal_ascii = [%sedlex.regexp? "0o", Plus '0' .. '7']

let hex_ascii =
  [%sedlex.regexp? "0x", Plus ('0' .. '9' | 'a' .. 'f' | 'A' .. 'F')]

let rec nom buf = match%sedlex buf with Plus any_blank -> nom buf | _ -> ()

let string buf =
  let buffer = Buffer.create 10 in
  let rec read_string buf =
    match%sedlex buf with
      | {|\"|} ->
          Buffer.add_char buffer '"';
          read_string buf
      | '"' -> STRING (Buffer.contents buffer)
      | Star (Compl '"') ->
          Buffer.add_string buffer (Utf8.lexeme buf);
          read_string buf
      | _ -> assert false
  in
  read_string buf

let digit_value c =
  let open Stdlib in
  match c with
    | 'a' .. 'f' -> 10 + Char.code c - Char.code 'a'
    | 'A' .. 'F' -> 10 + Char.code c - Char.code 'A'
    | '0' .. '9' -> Char.code c - Char.code '0'
    | _ -> assert false

let num_value buffer ~base ~first =
  let buf = Utf8.lexeme buffer in
  let c = ref 0 in
  for i = first to String.length buf - 1 do
    let v = digit_value buf.[i] in
    assert (v < base);
    c := (base * !c) + v
  done;
  ignore !c;
  buf

let token buf =
  nom buf;
  match%sedlex buf with
    | eof -> EOF
    | "" -> EOF
    | '-' -> MINUS
    | '+' -> PLUS
    | '"' -> string buf
    | ':' -> COLON
    | '[' -> LSQUARE
    | ']' -> RSQUARE
    | '{' -> LBRACKET
    | '}' -> RBRACKET
    | ',' -> COMMA
    | "true" -> BOOL true
    | "false" -> BOOL false
    | hex_ascii ->
        let number = num_value ~base:16 ~first:2 buf in
        NUMBER number
    | octal_ascii ->
        let number = num_value ~base:8 ~first:2 buf in
        NUMBER number
    | decimal_ascii ->
        let number = num_value ~base:10 ~first:0 buf in
        NUMBER number
    | _ ->
        let position = fst @@ lexing_positions buf in
        let tok = Utf8.lexeme buf in
        raise
        @@ LexError (position, Printf.sprintf "unexpected character %S" tok)

let rec read_all buf =
  match token buf with
    | EOF -> Printf.eprintf "DONE\n"
    | tok ->
        ignore tok;
        read_all buf

let read_whole_file filename =
  let ch = open_in_bin filename in
  let s = really_input_string ch (in_channel_length ch) in
  close_in ch;
  s

let () =
  let s = read_whole_file "/tmp/medialib.json" in
  let lexbuf = Sedlexing.Utf8.from_string s in
  read_all lexbuf
```
</details>