@hhugo 2023-06-11
I'm happy for this to be merged eventually but I still think we should add some tests.

Also a few comments:

- I don't think the PR is enough to really address the issue mentioned in the description. With the proposed implementation `Malformed` will be raised during refill, which will not allow you to have the location of the error.

- Why fix utf8 and not utf16 ?

- `Sedlexing.from_int_array` has the same issue..