@toots 2023-06-14
> I'm happy for this to be merged eventually but I still think we should add some tests.

There are a bunch of tests in `test/basic.ml` that seem designed to catch regression on all sort of lexbuf creation functions for `latin1`, `utf8` and `utf16`. What additional tests do you have in mind?

> Also a few comments:
> 
> * I don't think the PR is enough to really address the issue mentioned in the description. With the proposed implementation `Malformed` will be raised during refill, which will not allow you to have the location of the error.

Description clarified.

> * Why fix utf8 and not utf16 ?

Done, thanks for pointing out.

> * `Sedlexing.from_int_array` has the same issue..

I'm not sure that I see the issue:

```ocaml
let from_int_array a =
  let len = Array.length a in
  {
    empty_lexbuf with
    buf = Array.init len (fun i -> Uchar.of_int a.(i));
    len;
    finished = true;
  }
```

The original array doesn't seem copied here to me. And it's already in memory at call time. Pretty similar to having the string in memory at call time and peeling off of it during lexing.