@hhugo 2023-06-14
The issue with `from_int_array` is that `Uchar.of_int` can raise if the integer is not a valid codepoint.