@hhugo 2023-06-14
> There are a bunch of tests in test/basic.ml that seem designed to catch regression on all sort of lexbuf creation functions for latin1, utf8 and utf16. What additional tests do you have in mind?

I thing we should add tests where we partially lex the content with a malformed content on the part of the input that is not lexed.
(e.g. lexing an integer and a space from `123 \255\255`)