@toots 2023-06-14
While I think these are nice, this is also out of the scope of this PR. If you prefer I can remove all reference to `MalFormed` and merge this strictly as a performance optimisation, which is what I'm looking at with these changes. Better handling of `MalFormed` should be handled in your refactor.