@hhugo 2023-06-14
> With these changes, MalFormed errors might still be thrown while filling up the initial buffer when calling from_string but at least some errors that appears later in the string content will properly be raised during lexing.

- I don't think the buffer it filled during lexbuf creation.
- Malformed error raised during lexing will be done while refilling the buffer. The location wont be accurate.

Have you seen my reply on from_int_array ?

> The issue with `from_int_array` is that `Uchar.of_int` can raise if the integer is not a valid codepoint.