@toots 2023-06-23
Hi. Sorry I got busy.

I would like to merge this PR as a _performance improvement_. I have update the PR description. Do you have any pending concern about performances and/or regressions with it?